main_app/serializers.py
from rest_framework import serializers

from main_app.models import Publisher, BookInstance


class PublisherSerializer(serializers.ModelSerializer):
    class Meta:
        model = Publisher
        fields = ('id', 'name')


class BookInstanceSerializer(serializers.ModelSerializer):
    class Meta:
        model = BookInstance
        fields = ('id', 'title', 'is_sold')
